Updated Delete Row functionality
Before it was ambiguous. Say user wants to remover row A but selects row B and press "remove" button of row A then row B would be deleted..

Without js, it's unknown to me for now as how to make sure same button and corresponding selection was made in html & thymeleaf, so changed layout and functionality.